requirements.txt
python-gnupg==0.4.5
fs==2.4.11